+++ datawise-frontend/src/i18n/locales/en-US/dashboard.ts
@@ -53,6 +53,7 @@ export default {
         table: 'Table',
         connection: 'Connection',
         terminal: 'Terminal',
+        data_catalog: 'Data catalog',
     },
     status: {
         ok: 'Healthy',